requirements.txt
fastapi~=0.110.0
uvicorn[standard]~=0.27.1
aiogram~=3.20.0
asyncpg~=0.29.0
redis~=5.0.1
SQLAlchemy~=2.0.23
alembic~=1.12.1
argon2-cffi~=23.1.0
python-dotenv~=1.0.1
psycopg2-binary~=2.9.9
pytest~=8.1.1
httpx~=0.27.0
aiosqlite~=0.19.0
